requirements.txt
pytest~=7.4.3
Appium-Python-Client~=2.0.0
PyYAML~=6.0.1
selenium~=4.0.0
allure-pytest~=2.8.40
requests~=2.31.0
pytest-xdist~=3.4.0

